WIP security updates
