+++ ckanext/security/plugin/flask_plugin.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+import ckan.plugins as p
+import ckanext.security.views as views
+
+
+class MixinPlugin(p.SingletonPlugin):
+    p.implements(p.IBlueprint)
+
+    # IBlueprint
+
+    def get_blueprint(self):
+        return views.get_blueprints()
